requirements.txt
Flask
bokeh
gitpython
pybtex
python-dateutil
git+https://github.com/mrakitin/weather
